+++ src/DivideAndConquer/DivideAndConquer_BinarySearch.h
@@ -13,18 +13,18 @@ namespace mm {
 	//ObjectType should have comparison operators == and < defined for this Object type
 	//ContainerType should have operator[] defined to access element at any index ...preferably in O(1)
 	// and size() member function to give total number of elements
-	int binarySearch_1(
+	size_t binarySearch_1(
 		const ContainerType& container, 
 		const ObjectType& target, 
-		int rangeStart = 0,
-		int rangeEnd = -1)
+		size_t rangeStart = 0,
+		size_t rangeEnd = -1)
 	{
-		int start = rangeStart;
-		int end = rangeEnd != -1 ? rangeEnd : container.size() - 1;
+		size_t start = rangeStart;
+		size_t end = rangeEnd != -1 ? rangeEnd : container.size() - 1;
 
 		while (start <= end)
 		{
-			int mid = start + (end - start) / 2; //Do (start + end)/2 to avoid overflow
+			size_t mid = start + (end - start) / 2; //Do (start + end)/2 to avoid overflow
 			if (target == container[mid])
 				return mid;
 			if (target < container[mid])
