+++ src/DynamicProgramming/DP_HouseRobbery_UnitTests.cpp
@@ -69,7 +69,7 @@ namespace mm {
 
 		//cout << "\n\nProblem: HouseRobbery\nTrying sets of size 1 to 20...";
 		//Generate sets of sizes from 1 to 20
-		srand(time(nullptr));
+		srand(static_cast<unsigned int>(time(nullptr)));
 		for (int i = 1; i <= numTestCases; i++)
 		{
 			//Generate random treasury
@@ -84,7 +84,7 @@ namespace mm {
 
 		//cout << "\n\nTrying sets of size 20...";
 		//Generate various sets of size 20
-		srand(time(nullptr));
+		srand(static_cast<unsigned int>(time(nullptr)));
 		int size = 20;
 		treasurySet.resize(size);
 		for (int i = 0; i < numTestCases; i++)
